Hodge podge of stuff but should serve as a POC
