+++ .env-example
@@ -3,3 +3,4 @@ BASE_IMAGE_TYPE="base"
 OS_VERSION="ubuntu22.04"
 DATASET_URL="https://www.kaggle.com/api/v1/datasets/download/wenewone/cub2002011"
 SEED=123 # DO NOT CHANGE if you want to replicate results
+NO_ALBUMENTATIONS_UPDATE=1
