+++ aml/main.py
@@ -1,6 +1,6 @@
 import logging
 import sys
-from ViT.ViT_utils import optimize_hyperparameters, train_vit
+from ViT.ViT_utils import eval_vit, optimize_hyperparameters, train_vit
 from make_visualisations import make_visualization
 from make_labels import make_labels
 from random_forests.forest_train_util import train_classifier_forest, train_regressor_forest
@@ -26,6 +26,7 @@ command_map = {
     "--make_labels": (make_labels, "Creates a labels.csv file in /data"),
     "--make_visualization": (make_visualization, "Create visualizations of /data"),
     "--train_vit": (train_vit, "Train ViT and save the model to /data"),
+    "--eval_vit": (eval_vit, "Eval ViT"),
     "--optimize_hyperparams": (optimize_hyperparameters, "Optimize hyperparameters for ViT"),
     "--forest_regressor": (train_regressor_forest, "Train random forest regressor"),
     "--forest_classifier": (train_classifier_forest, "Train random forest classifier"),
