+++ aml/preprocessing/TreeImageDataset.py
@@ -28,7 +28,7 @@ class TreeImageDataset(Dataset):
         Raises:
             RuntimeError: Invalid dataset type.
         """
-        self._labels = pd.read_csv("/data/labels.csv", dtype=d_type, usecols=use_cols)
+        self._labels = pd.read_csv("/data/labels.csv", dtype=d_type, usecols=use_cols).iloc[:10]
         if type == "eval":
             self._base_transform = TreePrerocessPipeline.get_base_eval_transform()
         elif type == "train":
@@ -73,7 +73,7 @@ class TreeImageDataset(Dataset):
 
         return transformed_image, labels
 
-    def __getitem__(self, idx: int) -> tuple[NDArray, LabelType]:
+    def __getitem__(self, idx: int) -> tuple[torch.Tensor, LabelType]:
         """
         Retrieves image and label data for a given index.
 
@@ -119,3 +119,6 @@ class TreeImageDataset(Dataset):
             raise RuntimeError(f"Error during hog generation at index {idx}: {e}")
 
         return image_features, labels
+
+    def get_cls_labels(self) -> list[int]:
+        return [int(label) for label in self._labels["class_id"]]
